+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Translations: [简体中文](CHANGELOG.zh.md)
 
 * fix: Fixed the bug where the initial offset is incorrect when the content size is larger than the
   container size in read mode
+* improve: Improve the minScale algorithm, and now initialScale is smaller than baseScale or the
+  difference between the two is too small.
 * depend: Update kotlin to 2.1.10
 * depend: Update compose multiplatform to 1.8.0
 * depend: Update exifinterface to 1.4.1